requirements.txt
# Dependencias del Agente Comercial B2B

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0

# HTTP y APIs
requests>=2.31.0
httpx>=0.24.0

# PDF Processing
pypdf>=3.0.0
Pillow>=10.0.0

# OCR (opcional, para PDFs con imágenes)
pdf2image>=1.17.0
pytesseract>=0.3.10

# Vector Database / RAG
# Descomentar según implementación:
# pinecone-client>=2.2.0
# weaviate-client>=3.20.0
# chromadb>=0.4.0

# LLM / AI (para migración)
# openai>=1.0.0  # Para Llama vía OpenRouter/Together AI
# anthropic>=0.8.0  # Para Claude 4.5 Sonnet

# Base de datos
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0  # PostgreSQL

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging y Monitoreo
structlog>=23.1.0
sentry-sdk>=1.30.0

# Utilities
python-dateutil>=2.8.0
